# Conflicts:
#	app/calls/resources/classes/follow_me.php
#	app/dialplans/resources/switch/conf/dialplan/890_local_extension.xml
#	app/tftp/app_defaults.php
#	resources/install/scripts/app/ring_groups/index.lua
#	resources/install/scripts/app/xml_handler/resources/scripts/directory/directory.lua
#	resources/install/scripts/resources/functions/database.lua

+++ app/voicemails/voicemail_edit.php
@@ -23,16 +23,20 @@
  Contributor(s):
  Mark J Crane <markjcrane@fusionpbx.com>
 */
-require_once "root.php";
-require_once "resources/require.php";
-require_once "resources/check_auth.php";
-if (permission_exists('voicemail_add') || permission_exists('voicemail_edit')) {
-	//access granted
-}
-else {
-	echo "access denied";
-	exit;
-}
+
+//includes
+	require_once "root.php";
+	require_once "resources/require.php";
+	require_once "resources/check_auth.php";
+
+//check permissions
+	if (permission_exists('voicemail_add') || permission_exists('voicemail_edit')) {
+		//access granted
+	}
+	else {
+		echo "access denied";
+		exit;
+	}
 
 //add multi-lingual support
 	$language = new text;
@@ -59,14 +63,16 @@ else {
 			$voicemail_alternate_greet_id = check_str($_POST["voicemail_alternate_greet_id"]);
 			$voicemail_mail_to = check_str($_POST["voicemail_mail_to"]);
 			$voicemail_sms_to = check_str($_POST["voicemail_sms_to"]);
+			$voicemail_transcription_enabled = check_str($_POST["voicemail_transcription_enabled"]);
 			$voicemail_file = check_str($_POST["voicemail_file"]);
 			$voicemail_local_after_email = check_str($_POST["voicemail_local_after_email"]);
 			$voicemail_enabled = check_str($_POST["voicemail_enabled"]);
 			$voicemail_description = check_str($_POST["voicemail_description"]);
+			$voicemail_tutorial = check_str($_POST["voicemail_tutorial"]);
 		//remove the space
 			$voicemail_mail_to = str_replace(" ", "", $voicemail_mail_to);
-
-		echo "<pre>"; print_r($voicemail_options); echo "</pre>";
+		//debug info
+			//echo "<pre>"; print_r($voicemail_options); echo "</pre>";
 	}
 
 //unassign the voicemail id copy from the voicemail id
@@ -82,7 +88,7 @@ else {
 					d.voicemail_uuid = '".$voicemail_uuid."'";
 			$db->exec(check_sql($sqld));
 		//redirect the browser
-			$_SESSION["message"] = $text['message-delete'];
+			messages::add($text['message-delete']);
 			header("Location: voicemail_edit.php?id=".$voicemail_uuid);
 			return;
 	}
@@ -110,7 +116,7 @@ else {
 				)";
 			$db->exec(check_sql($sqli));
 		//redirect the browser
-			$_SESSION["message"] = $text['message-add'];
+			messages::add($text['message-add']);
 	}
 
 if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
@@ -147,8 +153,12 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				$sql .= "voicemail_alternate_greet_id, ";
 				$sql .= "voicemail_mail_to, ";
 				$sql .= "voicemail_sms_to, ";
+				$sql .= "voicemail_transcription_enabled, ";
+				$sql .= "voicemail_tutorial, ";
 				$sql .= "voicemail_file, ";
-				$sql .= "voicemail_local_after_email, ";
+				if (permission_exists('voicemail_local_after_email')) {
+					$sql .= "voicemail_local_after_email, ";
+				}
 				$sql .= "voicemail_enabled, ";
 				$sql .= "voicemail_description ";
 				$sql .= ")";
@@ -162,15 +172,19 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				$sql .= (($voicemail_alternate_greet_id != '') ? "'".$voicemail_alternate_greet_id."'" : 'null').", ";
 				$sql .= "'".$voicemail_mail_to."', ";
 				$sql .= "'".$voicemail_sms_to."', ";
+				$sql .= "'".$voicemail_transcription_enabled."', ";
+				$sql .= "'".$voicemail_tutorial."', ";
 				$sql .= "'".$voicemail_file."', ";
-				$sql .= "'".$voicemail_local_after_email."', ";
+				if (permission_exists('voicemail_local_after_email')) {
+					$sql .= "'".$voicemail_local_after_email."', ";
+				}
 				$sql .= "'".$voicemail_enabled."', ";
 				$sql .= "'".$voicemail_description."' ";
 				$sql .= ")";
 				$db->exec(check_sql($sql));
 				unset($sql);
 
-				$_SESSION["message"] = $text['message-add'];
+				messages::add($text['message-add']);
 			} //if ($action == "add")
 
 			if ($action == "update" && permission_exists('voicemail_edit')) {
@@ -181,8 +195,12 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				$sql .= "voicemail_alternate_greet_id = ".(($voicemail_alternate_greet_id != '') ? "'".$voicemail_alternate_greet_id."'" : 'null').", ";
 				$sql .= "voicemail_mail_to = '".$voicemail_mail_to."', ";
 				$sql .= "voicemail_sms_to = '".$voicemail_sms_to."', ";
+				$sql .= "voicemail_transcription_enabled = '".$voicemail_transcription_enabled."', ";
+				$sql .= "voicemail_tutorial = '".$voicemail_tutorial."', ";
 				$sql .= "voicemail_file = '".$voicemail_file."', ";
-				$sql .= "voicemail_local_after_email = '".$voicemail_local_after_email."', ";
+				if (permission_exists('voicemail_local_after_email')) {
+					$sql .= "voicemail_local_after_email = '".$voicemail_local_after_email."', ";
+				}
 				$sql .= "voicemail_enabled = '".$voicemail_enabled."', ";
 				$sql .= "voicemail_description = '".$voicemail_description."' ";
 				$sql .= "where domain_uuid = '".$domain_uuid."' ";
@@ -190,7 +208,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				$db->exec(check_sql($sql));
 				unset($sql);
 
-				$_SESSION["message"] = $text['message-update'];
+				messages::add($text['message-update']);
 			} //if ($action == "update")
 
 
@@ -214,11 +232,22 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 					$sql .= ") ";
 					$sql .= "values ";
 					foreach ($voicemail_options as $index => $voicemail_option) {
+
+						//set the uuid
 						$voicemail_option_uuid = uuid();
-						//seperate the action and the param
-						$option_array = explode(":", $voicemail_option["voicemail_option_param"]);
-						$voicemail_option['voicemail_option_action'] = array_shift($option_array);
-						$voicemail_option['voicemail_option_param'] = join(':', $option_array);
+
+						if (is_numeric($voicemail_option["voicemail_option_param"])) {
+							//if numeric then add tranfer $1 XML domain_name
+							$voicemail_option['voicemail_option_action'] = "menu-exec-app";
+							$voicemail_option['voicemail_option_param'] = "transfer ".$voicemail_option["voicemail_option_param"]." XML ".$_SESSION['domain_name'];
+						}
+						else {
+							//seperate the action and the param
+							$option_array = explode(":", $voicemail_option["voicemail_option_param"]);
+							$voicemail_option['voicemail_option_action'] = array_shift($option_array);
+							$voicemail_option['voicemail_option_param'] = join(':', $option_array);
+						}
+
 						//continue building insert query
 						$sql_record[$index] = "( ";
 						$sql_record[$index] .= "'".$voicemail_option_uuid."', ";
@@ -267,6 +296,8 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 			$voicemail_alternate_greet_id = $row["voicemail_alternate_greet_id"];
 			$voicemail_mail_to = $row["voicemail_mail_to"];
 			$voicemail_sms_to = $row["voicemail_sms_to"];
+			$voicemail_transcription_enabled = $row["voicemail_transcription_enabled"];
+			$voicemail_tutorial = $row["voicemail_tutorial"];
 			$voicemail_file = $row["voicemail_file"];
 			$voicemail_local_after_email = $row["voicemail_local_after_email"];
 			$voicemail_enabled = $row["voicemail_enabled"];
@@ -286,6 +317,8 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 //set defaults
 	if (strlen($voicemail_local_after_email) == 0) { $voicemail_local_after_email = "true"; }
 	if (strlen($voicemail_enabled) == 0) { $voicemail_enabled = "true"; }
+	if (strlen($voicemail_transcription_enabled) == 0) { $voicemail_transcription_enabled = "false"; }	
+	if (strlen($voicemail_tutorial) == 0) { $voicemail_tutorial = "false"; }
 
 //get the greetings list
 	$sql = "select * from v_voicemail_greetings ";
@@ -302,6 +335,39 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	require_once "resources/header.php";
 	$document['title'] = $text['title-voicemail'];
 
+//password complexity
+	$password_complexity = $_SESSION['voicemail']['password_complexity']['boolean'];
+	if ($password_complexity == "true") {
+		echo "<script>\n";	
+		$req['length'] = $_SESSION['voicemail']['password_min_length']['numeric'];
+		echo "	function check_password_strength(pwd) {\n";
+		echo "		var msg_errors = [];\n";
+		//length
+		if (is_numeric($req['length']) && $req['length'] != 0) {
+			echo "	var re = /.{".$req['length'].",}/;\n"; 
+			echo "	if (!re.test(pwd)) { msg_errors.push('".$req['length']."+ ".$text['label-digits']."'); }\n";
+		}
+		//numberic only
+		echo "		var re = /(?=.*[a-zA-Z\W])/;\n";
+		echo "		if (re.test(pwd)) { msg_errors.push('".$text['label-numberic_only']."'); }\n";
+		//repeating digits
+		echo "		var re = /(\d)\\1{2}/;\n";
+		echo "		if (re.test(pwd)) { msg_errors.push('".$text['label-password_repeating']."'); }\n";
+		//sequential digits
+		echo "		var re = /(012|123|345|456|567|678|789|987|876|765|654|543|432|321|210)/;\n";
+		echo "		if (re.test(pwd)) { msg_errors.push('".$text['label-password_sequential']."'); }\n";
+	
+		echo "		if (msg_errors.length > 0) {\n";
+		echo "			var msg = '".$text['message-password_requirements'].": ' + msg_errors.join(', ');\n";
+		echo "			display_message(msg, 'negative', '6000');\n";
+		echo "			return false;\n";
+		echo "		}\n";
+		echo "		else {\n";
+		echo "			return true;\n";
+		echo "		}\n";
+		echo "	}\n";
+		echo "</script>\n";
+	}
 //show the content
 	echo "<form method='post' name='frm' id='frm' action=''>\n";
 	echo "<table width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
@@ -312,7 +378,11 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "</td>\n";
 	echo "<td width='70%' align='right' valign='top'>\n";
 	echo "	<input type='button' class='btn' name='' alt='".$text['button-back']."' onclick=\"javascript:history.back();\" value='".$text['button-back']."'>\n";
-	echo "	<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
+	if ($password_complexity == "true") {
+		echo "		<input type='button' class='btn' value='".$text['button-save']."' onclick=\"if (check_password_strength(document.getElementById('password').value)) { submit_form(); }\">";
+	} else {
+		echo "	<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
+	}
 	echo "</td>\n";
 	echo "</tr>\n";
 
@@ -338,6 +408,20 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "</td>\n";
 	echo "</tr>\n";
 
+	echo "<tr>\n";
+	echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
+	echo "	".$text['label-voicemail_tutorial']."\n";
+	echo "</td>\n";
+	echo "<td class='vtable' align='left'>\n";
+	echo "	<select class='formfld' name='voicemail_tutorial' id='voicemail_tutorial'>\n";
+	echo "    	<option value='true' ".(($voicemail_tutorial == "true") ? "selected='selected'" : null).">".$text['label-true']."</option>\n";
+	echo "    	<option value='false' ".(($voicemail_tutorial == "false") ? "selected='selected'" : null).">".$text['label-false']."</option>\n";
+	echo "	</select>\n";
+	echo "<br />\n";
+	echo $text['description-voicemail_tutorial']."\n";
+	echo "</td>\n";
+	echo "</tr>\n";
+	
 	echo "<tr>\n";
 	echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
 	echo "	".$text['label-greeting']."\n";
@@ -490,6 +574,21 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 		echo "</td>\n";
 		echo "</tr>\n";
 	}
+	if(permission_exists('voicemail_transcription_edit') && $_SESSION['voicemail']['transcribe_enabled']['boolean'] == "true") {
+		echo "<tr>\n";
+		echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
+		echo "	".$text['label-voicemail_transcription_enabled']."\n";
+		echo "</td>\n";
+		echo "<td class='vtable' align='left'>\n";
+		echo "	<select class='formfld' name='voicemail_transcription_enabled' id='voicemail_transcription_enabled'>\n";
+		echo "    	<option value='true' ".(($voicemail_transcription_enabled == "true") ? "selected='selected'" : null).">".$text['label-true']."</option>\n";
+		echo "    	<option value='false' ".(($voicemail_transcription_enabled == "false") ? "selected='selected'" : null).">".$text['label-false']."</option>\n";
+		echo "	</select>\n";
+		echo "<br />\n";
+		echo $text['description-voicemail_transcription_enabled']."\n";
+		echo "</td>\n";
+		echo "</tr>\n";
+	}
 
 	echo "<tr>\n";
 	echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
@@ -506,19 +605,21 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "</td>\n";
 	echo "</tr>\n";
 
-	echo "<tr>\n";
-	echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
-	echo "	".$text['label-voicemail_local_after_email']."\n";
-	echo "</td>\n";
-	echo "<td class='vtable' align='left'>\n";
-	echo "	<select class='formfld' name='voicemail_local_after_email' id='voicemail_local_after_email' onchange=\"if (this.selectedIndex == 1) { document.getElementById('voicemail_file').selectedIndex = 2; }\">\n";
-	echo "    	<option value='true' ".(($voicemail_local_after_email == "true") ? "selected='selected'" : null).">".$text['label-true']."</option>\n";
-	echo "    	<option value='false' ".(($voicemail_local_after_email == "false") ? "selected='selected'" : null).">".$text['label-false']."</option>\n";
-	echo "	</select>\n";
-	echo "<br />\n";
-	echo $text['description-voicemail_local_after_email']."\n";
-	echo "</td>\n";
-	echo "</tr>\n";
+	if (permission_exists('voicemail_local_after_email')) {
+		echo "<tr>\n";
+		echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
+		echo "	".$text['label-voicemail_local_after_email']."\n";
+		echo "</td>\n";
+		echo "<td class='vtable' align='left'>\n";
+		echo "	<select class='formfld' name='voicemail_local_after_email' id='voicemail_local_after_email' onchange=\"if (this.selectedIndex == 1) { document.getElementById('voicemail_file').selectedIndex = 2; }\">\n";
+		echo "    	<option value='true' ".(($voicemail_local_after_email == "true") ? "selected='selected'" : null).">".$text['label-true']."</option>\n";
+		echo "    	<option value='false' ".(($voicemail_local_after_email == "false") ? "selected='selected'" : null).">".$text['label-false']."</option>\n";
+		echo "	</select>\n";
+		echo "<br />\n";
+		echo $text['description-voicemail_local_after_email']."\n";
+		echo "</td>\n";
+		echo "</tr>\n";
+	}
 
 	if ($action == "update") {
 		echo "	<tr>";
@@ -638,7 +739,11 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "				<input type='hidden' name='referer_path' value='".$http_referer['path']."'>\n";
 	echo "				<input type='hidden' name='referer_query' value='".$http_referer['query']."'>\n";
 	echo "				<br>";
-	echo "				<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
+	if ($password_complexity == "true") {
+		echo "			<input type='button' class='btn' value='".$text['button-save']."' onclick=\"if (check_password_strength(document.getElementById('password').value)) { submit_form(); }\">";
+	} else {
+		echo "			<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
+	}
 	echo "		</td>\n";
 	echo "	</tr>";
 	echo "</table>";
